move identifier
